Serialize ddsketch in trace stats snapshots
Serialize the sketches to dictionaries.

Right now sketches are stored in snapshots but are serialized to just
their length in bytes. This doesn't capture any meaningful information
about the sketches to ensure that they aren't changing.

The issue is that the buckets of the ddsketch are potentially quite noisy so
really the only information that can be captured and compared are the
settings of the ddsketch.