Add testing flag for snapshot generation (#79)
Previously, in order to generate snapshots the pytest fixture had to be
parametrized to be set to false. This was a bit clunky and not at all
intuitive for new developers.

Now a flag, `GENERATE_SNAPSHOTS=1` can be passed when running the
integration tests which will enable the snapshots to be output.