feat: add optional more lenient trace decoder (#122)
This change adds the ability to configure a more lenient trace decoder. The advantage of this decoder is that any floats/ints within the meta field will be left as strings. Additionally, long integers that were originally being decoded to a string are now decoded correctly as integers. To enable the optional tracer, set the DD_DISABLE_TRACE_PARSE_ERRORS or --disable-trace-parse-errors runtime flag with a value of True.


Co-authored-by: Kyle Verhoog <kyle@verhoog.ca>