Add `SNAPSHOT_REMOVED_ATTRS` to remove attributes from spans in snapshots (#125)
Add a new optional configuration to allow fields to be removed from the spans. By default, no keys will be removed from the snapshots.

💭 The motivation behind this feature is to get rid of some fields that can be noisy in the snapshots (e.g., start or duration) which could create some massive diffs when regenerating the snapshots.

Note that since the removal of the field start would cause an error when sorting the spans by this field, span_id was used as a fallback. Feel free to tell me if this would cause an issue I didn't consider (this is the reason why both start and span_id shouldn't be removed from the snapshots).