Ignore aiohttp warnings
aiohttp warns about using str keys on the app instance as it can
interfere with mypy typing. Right now it spams up the test logs and we
don't use it that heavily so ignore the warnings.

More on the warning: https://docs.aiohttp.org/en/stable/web_advanced.html#application-s-config