Fix UDS failing if socket already exists
In trying to use UDS in the Datadog injector test suite, I discovered
that the UDS socket functionality doesn't work particularly well in
containers. I suspect this is due to the atexit hook not being called
due to the abrupt container shutdown process. Subsequent runs of the
container will then fail as the file already exists and results in a
`OSError: [Errno 98] Address already in use` exception.