Fix container failing to start with UDS (#169)
As far as I can tell there is an issue with binding a socket and
`chmod`-ing it on macOS. I have repro'd the issue here:

```
>>> server_socket = socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
>>>
>>> server_socket.bind(socket_path)
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
NameError: name 'socket_path' is not defined
>>> server_socket.bind("/tmp/run/test.sock")
>>> import os
>>> os.chmod("/tmp/run/test.sock", 0o722)
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
OSError: [Errno 22] Invalid argument: '/tmp/run/test.sock'
```

I think it has something to due with volumes and sockets not playing
nicely together on docker for mac.

Anyway, this crashes the test agent on start-up and it's not the end of
the world if the permissions don't get set. So fail softly in this case.